requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
nltk==3.8.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# MLflow
mlflow==2.8.1

# Evidently
evidently==0.4.14

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
ruff==0.1.6
pre-commit==3.6.0

# Security
pip-audit==2.6.1

# Environment
python-dotenv==1.0.0